requirements.txt
python-telegram-bot==12.8
plotly==4.9.0
psutil==5.7.2
kaleido==0.0.3